+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Generalized support for setting (default) OIDC provider id through env var `OPENEO_AUTH_PROVIDER_ID`
   [#419](https://github.com/Open-EO/openeo-python-client/issues/419)
 - Added `OidcDeviceCodePollTimeout`: specific exception for OIDC device code flow poll timeouts
+- On-demand preview: Added `DataCube.preview()` to generate a XYZ service with the process graph and display a map widget
 
 ### Changed